Move both components into one file instead.
